develop: implemented calendar from library

+++ src/main/webapp/src/components/groups/UserGroupDetail.jsx
@@ -1,12 +1,12 @@
 import {Link, useLocation, useNavigate, useParams} from "react-router-dom";
-import Calendar from "../calendar/Calendar.jsx";
+import AppCalendar from "../calendar/AppCalendar.jsx";
 import {useEffect, useState} from "react";
 import {http} from "../../lib/http.jsx";
 import CalendarUser from "../users/CalendarUser.jsx";
-import {useAuth} from "../../features/AuthContext.jsx";
+import GroupsButton from "../buttons/GroupsButton.jsx";
 
 export default function UserGroupDetail() {
-    const { groupId } = useParams();
+    const {groupId} = useParams();
     const [groupMembers, setGroupMembers] = useState([])
     const [userRole, setUserRole] = useState("")
     const location = useLocation();
@@ -24,10 +24,11 @@ export default function UserGroupDetail() {
     async function getUserRole() {
         const res = await http.client.get(`groups/${groupId}/users/me/membershipRole`)
         setUserRole(res.data)
-        console.log(userRole)
     }
 
     async function leaveGroup() {
+        const ok = confirm(`Do you really want to leave ${groupName}?`)
+        if (!ok) return
         await http.client.delete(`groups/${groupId}/users/me`)
         navigate('/groups', {replace: true})
     }
@@ -38,19 +39,23 @@ export default function UserGroupDetail() {
     }, [groupId]);
 
     return <>
-        <Link to={"/groups"}>Groups</Link>
-        <h1>{groupName}</h1>
-        <button onClick={leaveGroup}>Leave Group</button>
-        {canManageMembership && (
-            <Link to={'manageMembers'}>Manage Members</Link>
-        )}
+        <header className={"sticky"}>
+            <nav>
+                <GroupsButton/>
+                <button className={"button"} onClick={leaveGroup}>Leave Group</button>
+                {canManageMembership && (
+                    <Link className={"button"} to={'manageMembers'}>Manage Members</Link>
+                )}
+            </nav>
+            <h1>{groupName}</h1>
+        </header>
         <ul>
-            {groupMembers.map(member => (
-                <li>
+            {groupMembers.map((member, i) => (
+                <li key={i}>
                     <CalendarUser user={member}/>
                 </li>
             ))}
         </ul>
-        <Calendar eventOwnerId={groupId} canAddEvents={canAddEvents}/>
+        <AppCalendar eventOwnerId={groupId} canAddEvents={canAddEvents}/>
     </>
 }
